# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JA4proxy Management UI
  description: Management interface for JA4proxy TLS security proxy
  version: 1.0.0
paths:
...
  /api/v1/config/reload:
    post:
      tags:
      - config
      summary: Reload Config
      description: 'Publish a config reload signal to all proxy instances.


        All instances subscribed to ``config.reload`` will hot-reload their

        configuration on receipt.'
      operationId: reload_config_api_v1_config_reload_post
      parameters:
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigReloadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/connections:
    get:
      tags:
      - connections
      summary: Get Connections
      description: 'Return recent connection events from the stream, optionally filtered.


        Reads up to limit*2 raw entries from the stream tail, applies filters,

        then returns at most *limit* results. ``truncated`` is True when the

        raw stream has more entries than were returned.'
      operationId: get_connections_api_v1_connections_get
      parameters:
      - name: ip
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by exact IP
          title: Ip
        description: Filter by exact IP
      - name: ja4
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by exact JA4 fingerprint
          title: Ja4
        description: Filter by exact JA4 fingerprint
      - name: action
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by action_taken
          title: Action
        description: Filter by action_taken
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: ISO 8601 timestamp — only events after this
          title: Since
        description: ISO 8601 timestamp — only events after this
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          description: Max results
          default: 100
          title: Limit
        description: Max results
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/fingerprints/{ja4}:
    get:
      tags:
      - connections
      summary: Get Fingerprint Detail
      description: 'Return aggregate statistics for a given JA4 fingerprint.


        Scans the full event stream for matching entries and returns total

        connection count, unique IPs, last-seen timestamp, and action breakdown.


        Returns 404 if the fingerprint has never appeared in the stream.'
      operationId: get_fingerprint_detail_api_v1_fingerprints__ja4__get
      parameters:
      - name: ja4
        in: path
        required: true
        schema:
          type: string
          title: Ja4
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/fingerprints/{ja4}/history:
    get:
      tags:
      - connections
      summary: Get Fingerprint History
      description: 'Return the chronological event history for a JA4 fingerprint.


        Reads the full stream and returns all matching events sorted oldest-first.'
      operationId: get_fingerprint_history_api_v1_fingerprints__ja4__history_get
      parameters:
      - name: ja4
        in: path
        required: true
        schema:
          type: string
          title: Ja4
      - name: token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ConfigReloadResponse:
      properties:
        message:
          type: string
          title: Message
          default: Config reload signal published
        published_to:
          type: string
          title: Published To
          default: config.reload
      type: object
      title: ConfigReloadResponse
      description: Response from POST /api/v1/config/reload.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...